+++ src/Common/InteractionHelper.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using Il2Cpp;
+using Il2CppInterop.Runtime;
 using MelonLoader;
 using UnityEngine;
 
@@ -68,6 +69,38 @@ internal class InteractionHelper
             buttonPromptLocId2: "No",
             confirmCallback: onConfirm,
             cancelCallback: null
+        );
+    }
+
+    /// <summary>Show a text input dialogue box which lets the player confirm or cancel.</summary>
+    /// <param name="question">The question text to display.</param>
+    /// <param name="initialText">The initial text in the input field.</param>
+    /// <param name="onConfirm">The action to perform with the input text when the player confirms.</param>
+    /// <param name="maxLength">The maximum input length.</param>
+    public void ShowTextInputDialogue(string question, string? initialText, Action<string> onConfirm, int maxLength = 80)
+    {
+        if (!this.TryGetUnusedConfirmationPanel(out Panel_Confirmation? panel))
+            return;
+
+        if (maxLength > 0)
+            panel.m_GenericMessageGroup.m_InputField.m_MaxLength = (uint)maxLength;
+
+        Panel_Confirmation.CallbackDelegate? confirmCallback = DelegateSupport.ConvertDelegate<Panel_Confirmation.CallbackDelegate>(
+            (Delegate)new Action(() => onConfirm(panel.m_GenericMessageGroup.m_InputField.GetText()))
+        );
+        if (confirmCallback is null)
+        {
+            this.Log.Warning("Can't show text input dialogue: failed to create confirmation callback.");
+            return;
+        }
+
+        panel.ShowRenamePanel(
+            question,
+            "Yes",
+            "No",
+            initialText ?? "",
+            confirmCallback,
+            null
         );
     }